move trampoline outside of engine.go where it is used
